Add `intl.error_level` into `config.neon`
Added `intl.error_level` configuration directive into `config.neon` to
throw error when `IntlDateFormatter::formatObject()` in `UserDataHelper`
fails. Otherwise it quietly returns false and we have no indication there
is something wrong.

We are not forcing it in ApplicationModule. This is PHP setting and
should be decided by developer.

remp/crm#2191