Update dependencies and configurations
- CRM modules are updated to version 2.1.

- Router translation resolver needs to be disabled by default
in order to migrate the project's database. Router resolver would
try to initialize routes of all of the modules possibly attempting
to read configuration from the database (like products module does).
That would fail because the DB would be empty.

It is safe to enable this back after the initial migration.

- Updated xdebug configuration to the Xdebug3 format.

- Added empty content folder required by exports of the CRM
modules.